User activeModification instead and set up map and dock resizing
Separate ModificationDock from Modification and move into Application
