+++ lib/actions/modifications.js
@@ -25,9 +25,7 @@ export const create = ({
     next: async (response) => {
       const modification = await response.json()
       return [
-        setLocally(modification),
-        setActive(modification),
-        toggleExpandedModification(modification)
+        setLocally(modification)
       ]
     }
   })
@@ -58,8 +56,6 @@ export const setAndRetrieveData = ({
   modification
 }) => [set(modification), getFeedsRoutesAndStops({bundleId, modifications: [modification]})]
 const setAll = createAction('set modifications')
-
-export const toggleExpandedModification = createAction('toggle expanded modification')
 
 export function getForScenario ({
   bundleId,
