+++ lib/application.js
@@ -15,6 +15,7 @@ import {EDIT_PROJECT_BOUNDS_COMPONENT} from './constants/map'
 import EditProjectBounds from './edit-project-bounds'
 import messages from './messages'
 import Modal from './components/modal'
+import ModificationDock from './modification-dock'
 import ScenarioMap from './scenario-map'
 
 import './map.css'
@@ -31,6 +32,7 @@ function mapStateToProps ({
   mapState,
   network,
   project,
+  scenario,
   user
 }, {
   location,
@@ -39,11 +41,13 @@ function mapStateToProps ({
   const {error, outstandingRequests} = network
   const hasError = !!error
   return {
+    activeModification: scenario ? scenario.activeModification : null,
     error: hasError ? error.error : '',
     errorMessage: hasError ? error.detailMessage : '',
     hasError,
     hasScenario: !!params.scenarioId,
     mapComponents: mapState.components || [],
+    modificationDockOffset: 0.25,
     outstandingRequests,
     project: project.projectsById[params.projectId],
     projects: project.projects,
@@ -85,7 +89,8 @@ class Application extends Component {
   }
 
   state = {
-    dockWidth: 0.25
+    dockWidth: 0.25,
+    modificationDockWidth: this.props.modificationDockOffset
   }
 
   componentWillMount () {
@@ -95,14 +100,36 @@ class Application extends Component {
   }
 
   onDockSizeChange = (dockWidth) => {
-    this.setState({dockWidth})
+    const newTotal = dockWidth + this.state.modificationDockWidth
+    if (newTotal <= 1 && dockWidth > 0) {
+      this.setState({dockWidth})
+    }
+  }
+
+  onModificationDockSizeChange = (modificationDockWidth) => {
+    // The second dock lives in a container that will be set to this size while
+    // the dock itself will be 100% of that container size. This made it possible to position relative to the first dock.
+    const newDockWidth = modificationDockWidth - this.props.modificationDockOffset
+    const newTotal = this.state.dockWidth + newDockWidth;
+
+    if (newTotal <= 1 && newDockWidth > 0) {
+      this.setState({modificationDockWidth: newDockWidth})
+    }
+    
+    return false
   }
 
   render () {
-    const {children, error, errorMessage, hasError, hasScenario, login, logout, mapComponents, outstandingRequests, projectId, project, userIsLoggedIn} = this.props
+    const {activeModification, children, error, errorMessage, hasError, 
+           hasScenario, login, logout, mapComponents,
+           outstandingRequests, projectId, project, userIsLoggedIn} = this.props
+
     const center = project != null
       ? latLng((project.bounds.north + project.bounds.south) / 2, (project.bounds.west + project.bounds.east) / 2)
       : DEFAULT_CENTER
+
+    const totalDockWidth = this.state.dockWidth + (activeModification ? this.state.modificationDockWidth : 0)
+    const mapWidth = (1 - totalDockWidth) * 100
 
     return (
       <div>
@@ -137,11 +164,33 @@ class Application extends Component {
             {children}
           </div>
         </Dock>
-
+        
+        {activeModification &&
+          <div className='modification-dock-container' style={{
+            width: `${this.state.modificationDockWidth * 100}%`,
+            position: 'fixed',
+            height: '100%',
+            top: '0',
+            left: `${this.state.dockWidth * 100}%`,
+            zIndex: '2499'
+          }}>
+            <Dock
+              dimMode='none'
+              fluid
+              isVisible
+              onSizeChange={this.onModificationDockSizeChange}
+              position='left'
+              size={1}
+              >
+              <ModificationDock modification={activeModification}></ModificationDock>
+            </Dock>
+          </div>
+        }
+        
         <div
           className='Fullscreen'
           style={{
-            width: `${(1 - this.state.dockWidth) * 100}%`
+            width: `${mapWidth}%`
           }}
           >
           <LeafletMap
