+++ lib/modification-group.js
@@ -47,7 +47,7 @@ export default class ModificationGroup extends DeepEqual {
     const shownModifications = modifications.slice(this.state.skip, this.state.limit)
     const hasMoreModifications = modifications.length > shownModifications.length
     const expanded = this.state.expanded && shownModifications.length !== 0
-    const iconName = expanded ? 'chevron-up' : 'chevron-down'
+    const iconName = expanded ? 'chevron-down' : 'chevron-up'
     const showOrHide = expanded ? 'Hide' : 'Show'
 
     return (
