+++ lib/reducers/scenario.js
@@ -109,15 +109,13 @@ export const reducers = {
     }
   },
   'set active modification' (state, action) {
-    const modification = action.payload
-    const isNotCurrentlyActive = !state.activeModification || !state.activeModification.id || state.activeModification.id !== modification.id
-    if (isNotCurrentlyActive) {
-      return {
-        ...state,
-        activeModification: {...modification}
-      }
-    } else {
-      return state
+    const modification = {...action.payload}
+    const activeModification = state.activeModification;
+    const newActive = (!activeModification || activeModification.id !== modification.id) ? modification : null
+
+    return {
+      ...state,
+      activeModification: newActive
     }
   },
   'set modification' (state, action) {
@@ -132,23 +130,6 @@ export const reducers = {
       modificationsById,
       modificationsByType: organizeByType(modifications)
     }
-  },
-  'toggle expanded modification' (state, action) {
-    const modification = {...action.payload}
-    const expandedModificationId = state.expandedModificationId
-    const modificationsById = {...state.modificationsById}
-    const newModifaction = modificationsById[modification.id]
-    let newState = {...state}
-    newModifaction.expanded = !modification.expanded
-    if (expandedModificationId && expandedModificationId !== modification.id) {
-      const oldModification = modificationsById[expandedModificationId]
-      oldModification.expanded = false
-      modificationsById[oldModification.id] = oldModification
-    }
-    const modifications = Object.values(modificationsById)
-    newState.modifications = modifications
-    newState.expandedModificationId = modification.id
-    return newState
   },
   'delete modification' (state, action) {
     const modificationId = action.payload
