+++ lib/scenario.js
@@ -69,7 +69,8 @@ function mapDispatchToProps (dispatch) {
     load: (id) => dispatch(load(id)),
     push: (opts) => dispatch(push(opts)),
     replaceModification: (opts) => dispatch(replaceModification(opts)),
-    runSinglePointAnalysis: (opts) => dispatch(runSinglePointAnalysis(opts))
+    runSinglePointAnalysis: (opts) => dispatch(runSinglePointAnalysis(opts)),
+    setActive: (modification) => dispatch(setActive(modification))
   }
 }
 
@@ -93,6 +94,7 @@ class Scenario extends Component {
     projectId: PropTypes.string.isRequired,
     replaceModification: PropTypes.func.isRequired,
     runSinglePointAnalysis: PropTypes.func.isRequired,
+    setActive: PropTypes.func.isRequired,
     variants: PropTypes.array.isRequired
   }
